+++ Pokelytics/app/src/main/AndroidManifest.xml
@@ -17,10 +17,10 @@
         tools:targetApi="31"
         android:name=".PokelyticsApp">
         <activity
-            android:name=".Splash"
+            android:name=".activities.SplashActivity"
             android:exported="false" />
         <activity
-            android:name=".Policy_eng"
+            android:name=".activities.PolicyActivity"
             android:exported="false" />
         <activity
             android:name=".activities.MainActivity"
