+++ Pokelytics/app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue_background"
-    tools:context=".Splash">
+    tools:context=".activities.SplashActivity">
 
 
     <ImageView
